+++ soffosai/src/common/serviceio_fields/ambiguity_detection_io.js
@@ -22,6 +22,7 @@ class AmbiguityDetectionIO extends ServiceIO {
       }
     };
   }
+
 }
 
 export default AmbiguityDetectionIO;