+++ soffosai/CHANGELOG.md
@@ -3,6 +3,11 @@
 To all who installed soffosai package lower than 0.0.4, please reinstall/update your package.
 
 
+## 0.1.4
+- added special validation on serviceio
+- added events: soffosai:on-request, soffosai:on-response, soffosai:on-service-error, soffosai:pipeline-start, soffosai:pipeline-end
+
+
 ## 0.1.3
 - inspect_arguments removed as it is not converting well during minification
 - soffosai.bundle.js created